+++ typings/src/classes/Utils.d.ts
@@ -1,8 +1,8 @@
 export declare class Utils {
     /**
-    * Checks for is the item object and returns it.
-    * @param {T} item The item to check.
-    * @returns {boolean} Is the item object or not.
-    */
+     * Checks for is the item object and returns it.
+     * @param {T} item The item to check.
+     * @returns {boolean} Is the item object or not.
+     */
     isObject<T>(item: T): boolean;
 }